requirements.txt
Flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
fpdf2==2.7.9
Werkzeug==3.0.1
